requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
argcomplete==3.6.2
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
IMAPClient==3.0.1
imgkit==1.2.3
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-barcode==0.15.1
python-dotenv==1.1.1
python-escpos==3.1
PyYAML==6.0.2
qrcode==8.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
